+++ src/server/mod.rs
@@ -5,9 +5,9 @@ mod dispatcher;
 mod errors;
 mod handshake;
 mod link;
-mod message;
 pub mod sasl;
 mod service;
+mod transfer;
 
 pub use self::app::App;
 pub use self::connect::{Connect, ConnectAck, ConnectOpened};
@@ -15,13 +15,16 @@ pub use self::control::{ControlFrame, ControlFrameKind};
 pub use self::errors::ServerError;
 pub use self::handshake::{handshake, Handshake};
 pub use self::link::Link;
-pub use self::message::{Message, Outcome};
 pub use self::sasl::Sasl;
 pub use self::service::Server;
+pub use self::transfer::{Outcome, Transfer};
 pub use crate::errors::{AmqpError, LinkError};
 pub use ntex_amqp_codec::protocol::Error;
 
 use crate::cell::Cell;
+
+#[doc(hidden)]
+pub type Message<T> = Transfer<T>;
 
 #[derive(Debug)]
 pub struct State<St>(Cell<St>);